refactor(react): create Tabs dir
