+++ packages/react/src/index.ts
@@ -24,7 +24,8 @@ export * from './components/ProgressBar.tsx'
 export * from './components/Radio.tsx'
 export * from './components/Select.tsx'
 export * from './components/Skeleton.tsx'
-export * from './components/tabs.tsx'
+// export * from './components/Tabs/tabs.tsx'
+export * from './components/Tabs/sections.tsx'
 export * from './components/Table.tsx'
 export * from './components/Tag.tsx'
 export * from './components/Textarea.tsx'
